Move common admin schemas to their own file
